Issue 리팩토링 + youtubeLive, media, issue 구조 변경 + issue 리팩토링
