+++ src/main/java/com/debateseason_backend_v1/domain/chatroom/model/response/chatroom/ResponseOnlyHome.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 import com.debateseason_backend_v1.domain.chatroom.model.response.chatroom.messages.Top5BestChatRoom;
 import com.debateseason_backend_v1.domain.chatroom.model.response.chatroom.type.ResponseWithTime;
-import com.debateseason_backend_v1.domain.issue.model.response.IssueBriefResponse;
+import com.debateseason_backend_v1.domain.issue.mapper.IssueRoomBriefMapper;
 import com.debateseason_backend_v1.domain.media.model.response.BreakingNewsResponse;
 
 import lombok.AllArgsConstructor;
@@ -22,7 +22,7 @@ public class ResponseOnlyHome {
 	private List<BreakingNewsResponse> breakingNews;
 
 	private List<Top5BestChatRoom> top5BestChatRooms;
-	private List<IssueBriefResponse> top5BestIssueRooms;
+	private List<IssueRoomBriefMapper> top5BestIssueRooms;
 
 	private List<ResponseWithTime> chatRoomResponse;
 }
