--- src/main/java/com/debateseason_backend_v1/domain/issue/PaginationDTO.java
@@ -1,14 +1,0 @@
-package com.debateseason_backend_v1.domain.issue;
-
-import java.util.List;
-
-import com.debateseason_backend_v1.domain.issue.model.response.IssueBriefResponse;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class PaginationDTO {
-	private List<IssueBriefResponse> items;
-}
