+++ src/main/java/com/debateseason_backend_v1/domain/repository/UserChatRoomRepository.java
@@ -84,6 +84,8 @@ public interface UserChatRoomRepository extends JpaRepository<UserChatRoom, Long
 	);
 
 
+	//Legacy
+	/*
 	// 2-1 이슈방 issue-id로만 조회
 	@Query(value = "SELECT chat_room_id FROM chat_room WHERE issue_id = :issueId ORDER BY chat_room_id DESC LIMIT 3", nativeQuery = true)
 	List<Long> findTop3ChatRoomIdsByIssueId(@Param("issueId") Long issueId);
@@ -94,6 +96,10 @@ public interface UserChatRoomRepository extends JpaRepository<UserChatRoom, Long
 		@Param("issueId") Long issueId,
 		@Param("ChatRoomId") Long ChatRoomId
 	);
+
+	 */
+
+	/*
 	@Query(value = """
     SELECT ch.chat_room_id, ch.title, ch.content, ch.created_at,
            COUNT(CASE WHEN ucr.opinion = 'AGREE' THEN 1 END) AS AGREE,
@@ -105,6 +111,8 @@ public interface UserChatRoomRepository extends JpaRepository<UserChatRoom, Long
     ORDER BY ch.chat_room_id DESC
     """, nativeQuery = true)
 	List<Object[]> findChatRoomAggregates(@Param("chatRoomIds") List<Long> chatRoomIds);
+
+	 */
 
 	@Query(value = """
     SELECT chat_room_id, opinion AS opinion
