+++ src/main/java/com/debateseason_backend_v1/domain/repository/UserIssueRepository.java
@@ -1,7 +1,7 @@
 package com.debateseason_backend_v1.domain.repository;
 
 
-import com.debateseason_backend_v1.domain.repository.entity.Issue;
+import com.debateseason_backend_v1.domain.issue.infrastructure.entity.IssueEntity;
 import com.debateseason_backend_v1.domain.user.infrastructure.UserEntity;
 import com.debateseason_backend_v1.domain.repository.entity.UserIssue;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -15,7 +15,7 @@ import java.util.List;
 public interface UserIssueRepository extends JpaRepository<UserIssue,Long> {
 
 
-    UserIssue findByIssueAndUser(Issue issue, UserEntity user);
+    UserIssue findByIssueEntityAndUser(IssueEntity issueEntity, UserEntity user);
 
     @Query(value = "SELECT bookmark FROM user_issue WHERE issue_id = :issueId AND user_id = :userId", nativeQuery = true)
     List<Object[]> findByIssueIdAndUserId(@Param("issueId") Long issueId, @Param("userId") Long userId);
