--- src/main/java/com/debateseason_backend_v1/domain/repository/entity/Issue.java
@@ -1,45 +1,0 @@
-package com.debateseason_backend_v1.domain.repository.entity;
-
-import java.time.LocalDateTime;
-
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-@Builder
-@Entity
-@EntityListeners(AuditingEntityListener.class) // 이걸 붙여야 @CreatedDate가 활성화되서 자동으로 날짜 입력
-public class Issue {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "issue_id")
-	private Long id;
-
-	private String title;
-
-	@Column(name = "major_category")
-	private String majorCategory;
-
-	private String middleCategory;
-
-	@CreatedDate
-	@Column(name = "created_at", updatable = false)
-	private LocalDateTime createdAt;
-
-}
