+++ src/main/java/com/debateseason_backend_v1/domain/repository/entity/UserIssue.java
@@ -1,5 +1,6 @@
 package com.debateseason_backend_v1.domain.repository.entity;
 
+import com.debateseason_backend_v1.domain.issue.infrastructure.entity.IssueEntity;
 import com.debateseason_backend_v1.domain.user.infrastructure.UserEntity;
 
 import jakarta.persistence.Entity;
@@ -33,7 +34,7 @@ public class UserIssue {
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "issue_id", nullable = false)
-	private Issue issue;
+	private IssueEntity issueEntity;
 
 	private String bookmark;
 }
